# Conflicts:
#	MIDAS/src/finite-state-machines/fsm_states.h
#	MIDAS/src/systems.cpp

+++ MIDAS/lib/ADS7138Q1/ads7138-q1.cpp
@@ -26,7 +26,7 @@ static AdcRegReadResult adc_reg_read(uint8_t address, uint8_t reg_address) {
   if(!WIRE.endTransmission(true)){
     return AdcRegReadResult{.value=0, .error=AdcError::I2CError};
   }
-  uint8_t ct = WIRE.requestFrom(address, 1, 1);
+  uint8_t ct = WIRE.requestFrom((int)address, 1, 1);
   if(ct != 1){
     return AdcRegReadResult{.value=0, .error=AdcError::I2CError};
   }
@@ -52,7 +52,7 @@ AdcReadResult adcAnalogRead(ADCAddress pin){
 
     adc_reg_write(ADC_ADDR, ADC_MANUAL_CH_SEL, pin.pin_id);
     
-    int ct = WIRE.requestFrom(ADC_ADDR, 2, 1);
+    int ct = WIRE.requestFrom((int)ADC_ADDR, 2, 1);
     if(ct != 2){
         return AdcReadResult{.value=0, .error=AdcError::I2CError};
     }
